pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "steghide-lab"
version = "0.1.0"
description = "LSB steganography tool for hiding and extracting messages in PNG images."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "Pillow>=10.0.0",
    "cryptography>=41.0.0",
]

[project.scripts]
steghide-lab = "cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["steganography*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
